🐛 kc 240428 订货本3.1 修改合计、下单判断
(cherry picked from commit 47a3e69dd0f6e7410838caef161f4c8ed46a8aed)

+++ src/main/java/com/base/sbc/module/orderbook/service/OrderBookDetailService.java
@@ -15,6 +15,7 @@ import com.github.pagehelper.PageInfo;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -65,7 +66,7 @@ public interface OrderBookDetailService extends BaseService<OrderBookDetail> {
      */
     boolean assignmentDesigner(List<OrderBookDetailSaveDto> dto);
 
-    Map<String, Double> queryCount(OrderBookDetailQueryDto dto);
+    Map<String, BigDecimal> queryCount(OrderBookDetailQueryDto dto);
 
     void submitForApproval(OrderBookDetailSaveDto dto);
